backend/pyproject.toml
[project]
name = "app"
version = "0.0.0"
description = ""
authors = [
    {name = "artem",email = "artemm469@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "pydantic (>=2.12.3,<3.0.0)",
    "fastapi (>=0.120.0,<0.121.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "app", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
